+++ src/ai/backend/manager/cli/image_impl.py
@@ -16,7 +16,8 @@ from ai.backend.common.exception import UnknownImageReference
 from ai.backend.common.types import ImageAlias
 from ai.backend.logging import BraceStyleAdapter
 
-from ..models.image import ImageAliasRow, ImageIdentifier, ImageRow, ImageStatus
+from ..data.image.types import ImageStatus
+from ..models.image import ImageAliasRow, ImageIdentifier, ImageRow
 from ..models.image import rescan_images as rescan_images_func
 from ..models.utils import connect_database
 from .context import CLIContext, redis_ctx
