+++ src/ai/backend/manager/models/gql_models/__init__.py
@@ -3,6 +3,7 @@ from . import group as _group
 from . import image as _image
 from . import kernel as _kernel
 from . import keypair as _keypair
+from . import resource_preset as _resource_preset
 from . import scaling_group as _scaling_group
 from . import session as _session
 from . import user as _user
@@ -13,6 +14,7 @@ __all__ = (
     *_image.__all__,
     *_kernel.__all__,
     *_keypair.__all__,
+    *_resource_preset.__all__,
     *_scaling_group.__all__,
     *_session.__all__,
     *_user.__all__,
@@ -23,6 +25,7 @@ from .group import *  # noqa
 from .image import *  # noqa
 from .kernel import *  # noqa
 from .keypair import *  # noqa
+from .resource_preset import *  # noqa
 from .scaling_group import *  # noqa
 from .session import *  # noqa
 from .user import *  # noqa
