+++ src/ai/backend/manager/models/image.py
@@ -40,7 +40,12 @@ from ai.backend.common.types import (
 )
 from ai.backend.common.utils import join_non_empty
 from ai.backend.logging import BraceStyleAdapter
-from ai.backend.manager.data.image.types import ImageLabelsData, ImageResourcesData
+from ai.backend.manager.data.image.types import (
+    ImageLabelsData,
+    ImageResourcesData,
+    ImageStatus,
+    ImageType,
+)
 
 from ..api.exceptions import ImageNotFound
 from ..container_registry import get_container_registry_cls
@@ -297,22 +302,11 @@ async def rescan_images(
     # TODO: delete images removed from registry?
 
 
-class ImageType(enum.Enum):
-    COMPUTE = "compute"
-    SYSTEM = "system"
-    SERVICE = "service"
-
-
 # Defined for avoiding circular import
 def _get_image_endpoint_join_condition():
     from ai.backend.manager.models.endpoint import EndpointRow
 
     return ImageRow.id == foreign(EndpointRow.image)
-
-
-class ImageStatus(enum.StrEnum):
-    ALIVE = "ALIVE"
-    DELETED = "DELETED"
 
 
 class ImageRow(Base):
