+++ src/ai/backend/manager/services/resource_preset/actions/create_preset.py
@@ -1,29 +1,10 @@
 from dataclasses import dataclass
-from typing import Any, Optional, override
+from typing import Optional, override
 
-from ai.backend.common.types import BinarySize, ResourceSlot
 from ai.backend.manager.actions.action import BaseActionResult
 from ai.backend.manager.models.resource_preset import ResourcePresetRow
 from ai.backend.manager.services.resource_preset.actions.base import ResourcePresetAction
-from ai.backend.manager.types import Creator
-
-
-@dataclass
-class ResourcePresetCreator(Creator):
-    name: str
-    resource_slots: ResourceSlot
-    shared_memory: Optional[str]
-    scaling_group_name: Optional[str]
-
-    def fields_to_store(self) -> dict[str, Any]:
-        return {
-            "name": self.name,
-            "resource_slots": self.resource_slots,
-            "shared_memory": BinarySize.from_str(self.shared_memory)
-            if self.shared_memory
-            else None,
-            "scaling_group_name": self.scaling_group_name,
-        }
+from ai.backend.manager.services.resource_preset.types import ResourcePresetCreator
 
 
 @dataclass
@@ -41,7 +22,7 @@ class CreateResourcePresetAction(ResourcePresetAction):
 
 @dataclass
 class CreateResourcePresetActionResult(BaseActionResult):
-    # TODO: Create ResourcePresetRow type
+    # TODO: Use dataclass instead of Row
     resource_preset: ResourcePresetRow
 
     @override
